requirements.txt
pytest==7.3.1
allure-pytest==2.13.2
selene==2.0.0rc2
iniconfig==2.0.0